app/requirements.txt
attrs==21.4.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mongoengine==0.23.1
numpy==1.21.3
opencv-python-headless==4.5.5.64
packaging==21.3
pdf2image==1.16.0
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pymongo==3.12.1
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.19.2
tomli==2.0.1
Werkzeug==2.0.2
WTForms==2.3.1
